+++ spec/shoulda/matchers/action_controller/strong_parameters_matcher_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe Shoulda::Matchers::ActionController do
-  describe "#permit" do
+  describe '#permit' do
     it 'matches when the sent parameter is allowed' do
       controller_for_resource_with_strong_parameters(action: :create) do
         params.require(:user).permit(:name)
@@ -56,8 +56,8 @@ describe Shoulda::Matchers::ActionController::StrongParametersMatcher do
     end
   end
 
-  describe "#matches?" do
-    it "is true for a subset of the allowable attributes" do
+  describe '#matches?' do
+    it 'is true for a subset of the allowable attributes' do
       controller_for_resource_with_strong_parameters(action: :create) do
         params.require(:user).permit(:name)
       end
@@ -66,7 +66,7 @@ describe Shoulda::Matchers::ActionController::StrongParametersMatcher do
       expect(matcher.matches?(@controller)).to be_true
     end
 
-    it "is true for all the allowable attributes" do
+    it 'is true for all the allowable attributes' do
       controller_for_resource_with_strong_parameters(action: :create) do
         params.require(:user).permit(:name, :age)
       end
@@ -75,7 +75,7 @@ describe Shoulda::Matchers::ActionController::StrongParametersMatcher do
       expect(matcher.matches?(@controller)).to be_true
     end
 
-    it "is false when any attributes are not allowed" do
+    it 'is false when any attributes are not allowed' do
       controller_for_resource_with_strong_parameters(action: :create) do
         params.require(:user).permit(:name)
       end
@@ -84,67 +84,130 @@ describe Shoulda::Matchers::ActionController::StrongParametersMatcher do
       expect(matcher.matches?(@controller)).to be_false
     end
 
-    it "is false when permit is not called" do
+    it 'is false when permit is not called' do
       controller_for_resource_with_strong_parameters(action: :create)
 
       matcher = described_class.new([:name]).in_context(self).for(:create)
       expect(matcher.matches?(@controller)).to be_false
     end
 
-    it "requires an action" do
+    it 'requires an action' do
       controller_for_resource_with_strong_parameters
       matcher = described_class.new([:name])
-      expect { matcher.matches?(@controller) }
-        .to raise_error(Shoulda::Matchers::ActionController::StrongParametersMatcher::ActionNotDefinedError)
+      expect { matcher.matches?(@controller) }.
+        to raise_error(described_class::ActionNotDefinedError)
     end
 
-    it "requires a verb for non-restful action" do
+    it 'requires a verb for non-restful action' do
       controller_for_resource_with_strong_parameters
       matcher = described_class.new([:name]).for(:authorize)
-      expect { matcher.matches?(@controller) }
-        .to raise_error(Shoulda::Matchers::ActionController::StrongParametersMatcher::VerbNotDefinedError)
+      expect { matcher.matches?(@controller) }.
+        to raise_error(described_class::VerbNotDefinedError)
+    end
+
+    it 'works with routes that require extra params' do
+      options = {
+        controller_name: 'Posts',
+        action: :show,
+        routes: -> {
+          get '/posts/:slug', to: 'posts#show'
+        }
+      }
+      controller_for_resource_with_strong_parameters(options) do
+        params.require(:user).permit(:name)
+      end
+
+      matcher = described_class.new([:name]).
+        in_context(self).
+        for(:show, verb: :get, params: { slug: 'foo' })
+      expect(matcher.matches?(@controller)).to be_true
+    end
+
+    it 'works with #update specifically' do
+      controller_for_resource_with_strong_parameters(action: :update) do
+        params.require(:user).permit(:name)
+      end
+
+      matcher = described_class.new([:name]).
+        in_context(self).
+        for(:update, params: { id: 1 })
+      expect(matcher.matches?(@controller)).to be_true
+    end
+
+    it 'does not raise an error when #fetch was used instead of #require (issue #495)' do
+      controller_for_resource_with_strong_parameters(action: :create) do
+        params.fetch(:order, {}).permit(:eta, :diner_id)
+      end
+
+      matcher = described_class.new([:eta, :diner_id]).
+        in_context(self).
+        for(:create)
+      expect(matcher.matches?(@controller)).to be_true
+    end
+
+    it 'tracks multiple calls to #permit' do
+      sets_of_attributes = [
+        [:eta, :diner_id],
+        [:phone_number, :address_1, :address_2, :city, :state, :zip]
+      ]
+      controller_for_resource_with_strong_parameters(action: :create) do
+        params.require(:order).permit(sets_of_attributes[0])
+        params.require(:diner).permit(sets_of_attributes[1])
+      end
+
+      matcher = described_class.new(sets_of_attributes[0]).
+        in_context(self).
+        for(:create)
+      expect(matcher.matches?(@controller)).to be_true
+
+      matcher = described_class.new(sets_of_attributes[1]).
+        in_context(self).
+        for(:create)
+      expect(matcher.matches?(@controller)).to be_true
     end
 
     context 'stubbing params on the controller' do
       it 'still allows the original params to be set and accessed' do
-        actual_value = nil
+        actual_user_params = nil
+        actual_foo_param = nil
 
         controller_for_resource_with_strong_parameters(action: :create) do
           params[:foo] = 'bar'
-          actual_value = params[:foo]
+          actual_foo_param = params[:foo]
+
+          actual_user_params = params[:user]
 
           params.require(:user).permit(:name)
         end
 
-        matcher = described_class.new([:name]).in_context(self).for(:create)
+        matcher = described_class.new([:name]).
+          in_context(self).
+          for(:create, params: { user: { some: 'params' } })
         matcher.matches?(@controller)
 
-        expect(actual_value).to eq 'bar'
+        expect(actual_user_params).to eq('some' => 'params')
+        expect(actual_foo_param).to eq 'bar'
       end
 
-      it 'stubs the params while the controller action is being run' do
-        params_class = nil
-
+      it 'stubs the params during the controller action' do
         controller_for_resource_with_strong_parameters(action: :create) do
-          params_class = params.class
-          params.require(:user).permit(:name)
+          params.require(:user)
         end
 
         matcher = described_class.new([:name]).in_context(self).for(:create)
-        matcher.matches?(@controller)
 
-        expect(params_class).to be described_class::StubbedParameters
+        expect { matcher.matches?(@controller) }.not_to raise_error
       end
 
       it 'does not permanently stub params' do
-        controller_for_resource_with_strong_parameters(action: :create) do
-          params.require(:user).permit(:name)
-        end
+        controller_for_resource_with_strong_parameters(action: :create)
 
         matcher = described_class.new([:name]).in_context(self).for(:create)
         matcher.matches?(@controller)
 
-        expect(@controller.params).to be_a(ActionController::Parameters)
+        expect {
+          @controller.params.require(:user)
+        }.to raise_error(::ActionController::ParameterMissing)
       end
 
       it 'prevents permanently stubbing params on error' do
@@ -156,64 +219,79 @@ describe Shoulda::Matchers::ActionController::StrongParametersMatcher do
         rescue SimulatedError
         end
 
-        expect(@controller.params).to be_a(ActionController::Parameters)
+        expect {
+          @controller.params.require(:user)
+        }.to raise_error(::ActionController::ParameterMissing)
       end
     end
   end
 
-  describe "failure message" do
-    it "includes all missing attributes" do
+  describe 'failure message' do
+    it 'includes all missing attributes' do
       controller_for_resource_with_strong_parameters(action: :create) do
         params.require(:user).permit(:name, :age)
       end
 
       expect {
         expect(@controller).to permit(:name, :age, :city, :country).for(:create)
-      }.to fail_with_message("Expected controller to permit city and country, but it did not.")
+      }.to fail_with_message('Expected controller to permit city and country, but it did not.')
     end
 
-    it "includes all attributes that should not have been allowed but were" do
+    it 'includes all attributes that should not have been allowed but were' do
       controller_for_resource_with_strong_parameters(action: :create) do
         params.require(:user).permit(:name, :age)
       end
 
       expect {
         expect(@controller).not_to permit(:name, :age).for(:create)
-      }.to fail_with_message("Expected controller not to permit name and age, but it did.")
+      }.to fail_with_message('Expected controller not to permit name and age, but it did.')
     end
   end
 
-  describe "#for" do
-    context "when given :create" do
-      it "posts to the controller" do
+  describe '#for' do
+    context 'when given :create' do
+      it 'POSTs to the controller' do
         controller = ActionController::Base.new
-        context = stub('context', post: nil)
+        context = mock()
+        context.expects(:post).with(:create, {})
         matcher = described_class.new([:name]).in_context(context).for(:create)
 
         matcher.matches?(controller)
-        expect(context).to have_received(:post).with(:create)
       end
     end
 
-    context "when given :update" do
-      it "puts to the controller" do
-        controller = ActionController::Base.new
-        context = stub('context', put: nil)
-        matcher = described_class.new([:name]).in_context(context).for(:update)
+    context 'when given :update' do
+      if rails_gte_41?
+        it 'PATCHes to the controller' do
+          controller = ActionController::Base.new
+          context = mock()
+          context.expects(:patch).with(:update, {})
+          matcher = described_class.new([:name]).in_context(context).for(:update)
 
-        matcher.matches?(controller)
-        expect(context).to have_received(:put).with(:update)
+          matcher.matches?(controller)
+        end
+      else
+        it 'PUTs to the controller' do
+          controller = ActionController::Base.new
+          context = mock()
+          context.expects(:put).with(:update, {})
+          matcher = described_class.new([:name]).in_context(context).for(:update)
+
+          matcher.matches?(controller)
+        end
       end
     end
 
-    context "when given a custom action and verb" do
-      it "deletes to the controller" do
+    context 'when given a custom action and verb' do
+      it 'calls the action with the verb' do
         controller = ActionController::Base.new
-        context = stub('context', delete: nil)
-        matcher = described_class.new([:name]).in_context(context).for(:hide, verb: :delete)
+        context = mock()
+        context.expects(:delete).with(:hide, {})
+        matcher = described_class.new([:name]).
+          in_context(context).
+          for(:hide, verb: :delete)
 
         matcher.matches?(controller)
-        expect(context).to have_received(:delete).with(:hide)
       end
     end
   end